# html/laravel/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Invillia PHP CHALLENGE OpenApi Documentation",
        "description": "PHP CHALLENGE - developed by Leandro Manzano Merlo",
        "contact": {
            "email": "leandro.merlo@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api/v1",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/shiporders": {
            "get": {
                "tags": [
                    "Ship Orders"
                ],
                "summary": "Get list of Ship Orders",
                "description": "Returns list of all Ship Orders",
                "operationId": "getShipOrdersList",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ShipOrdersResource"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
        "/shiporders/{id}": {
            "get": {
                "tags": [
                    "Ship Orders"
                ],
                "summary": "Find a Ship Order by id",
                "description": "Finds a Ship Order from a given id",
                "operationId": "findShipOrderById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Ship Order id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ShipOrder"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
        "/shiporders/{id}/shipto": {
            "get": {
                "tags": [
                    "Ship Orders"
                ],
                "summary": "Find a Ship To (Destination) by Ship Order id",
                "description": "Finds a Ship To (Destination) from a given Ship Order id",
                "operationId": "findShipToByShipOrderId",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Ship Order id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ShipTo"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
        "/shiporders/{id}/items": {
            "get": {
                "tags": [
                    "Ship Orders"
                ],
                "summary": "Find Items by Ship Order id",
                "description": "Finds a Items from a given Ship Order id",
                "operationId": "findItemsByShipOrderId",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Ship Order id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Item"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
        "/shiptos": {
            "get": {
                "tags": [
                    "Ship To (Destination)"
                ],
                "summary": "Get list of Ship To (Destinations)",
                "description": "Returns list of all Ship To (Destinations)",
                "operationId": "getShipTosList",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ShipTosResource"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
        "/shiptos/{id}": {
            "get": {
                "tags": [
                    "Ship To (Destination)"
                ],
                "summary": "Find a destination by id",
                "description": "Finds a destination from a given id",
                "operationId": "findShipToById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Ship To id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ShipTo"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
            "Item": {
                "title": "Item",
                "description": "Item model",
                "properties": {
                    "id": {
                        "title": "id",
                        "description": "ID",
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "title": {
                        "title": "title",
                        "description": "Item title",
                        "type": "string",
                        "example": "Title"
                    },
                    "quantity": {
                        "title": "quantity",
                        "description": "Quantity",
                        "type": "integer",
                        "format": "int",
                        "example": 20
                    },
                    "price": {
                        "title": "id",
                        "description": "ID",
                        "type": "number",
                        "format": "float",
                        "example": 754.34
                    },
                    "ship_order_id": {
                        "title": "ship_order_id",
                        "description": "Ship Order Id",
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "created_at": {
                        "title": "Created at",
                        "description": "Created at",
                        "type": "string",
                        "format": "datetime",
                        "example": "2020-01-27 17:50:45"
                    },
                    "updated_at": {
                        "title": "Updated at",
                        "description": "Updated at",
                        "type": "string",
                        "format": "datetime",
                        "example": "2020-01-27 17:50:45"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "Item"
                }
            },
...
            "ShipOrder": {
                "title": "ShipOrder",
                "description": "ShipOrder model",
                "properties": {
                    "id": {
                        "title": "id",
                        "description": "ID",
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "person_id": {
                        "title": "person_id",
                        "description": "Person Id",
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "ship_to_id": {
                        "title": "ship_to_id",
                        "description": "Ship To Id",
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "created_at": {
                        "title": "Created at",
                        "description": "Created at",
                        "type": "string",
                        "format": "datetime",
                        "example": "2020-01-27 17:50:45"
                    },
                    "updated_at": {
                        "title": "Updated at",
                        "description": "Updated at",
                        "type": "string",
                        "format": "datetime",
                        "example": "2020-01-27 17:50:45"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "ShipOrder"
                }
            },
            "ShipTo": {
                "title": "ShipTo",
                "description": "Ship To (Destination) model",
                "properties": {
                    "id": {
                        "title": "id",
                        "description": "ID",
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "name": {
                        "title": "name",
                        "description": "Ship to name",
                        "type": "string",
                        "example": "Name"
                    },
                    "address": {
                        "title": "address",
                        "description": "Ship to address",
                        "type": "string",
                        "example": "Address"
                    },
                    "city": {
                        "title": "ciry",
                        "description": "Ship to city",
                        "type": "string",
                        "example": "City"
                    },
                    "country": {
                        "title": "country",
                        "description": "Ship to country",
                        "type": "string",
                        "example": "Country"
                    },
                    "created_at": {
                        "title": "Created at",
                        "description": "Created at",
                        "type": "string",
                        "format": "datetime",
                        "example": "2020-01-27 17:50:45"
                    },
                    "updated_at": {
                        "title": "Updated at",
                        "description": "Updated at",
                        "type": "string",
                        "format": "datetime",
                        "example": "2020-01-27 17:50:45"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "ShipTo"
                }
            },
...
            "ShipOrdersResource": {
                "title": "ShipOrdersResource",
                "description": "ShipOrders resource",
                "properties": {
                    "data": {
                        "title": "List of Ship Ordersw",
                        "description": "Data wrapper",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/ShipOrder"
                        }
                    }
                },
                "type": "object",
                "xml": {
                    "name": "ShipOrdersResource"
                }
            },
            "ShipTosResource": {
                "title": "ShipTosResource",
                "description": "Ship To (Destination) resource",
                "properties": {
                    "data": {
                        "title": "List of Ship To (Destination) ",
                        "description": "Data wrapper",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/ShipTo"
                        }
                    }
                },
                "type": "object",
                "xml": {
                    "name": "ShipTosResource"
                }
            }
        },
        "securitySchemes": {
            "apiAuth": {
                "type": "http",
                "description": "Login with email and password to get the authentication token",
                "name": "Token based",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}